feat: add google recaptcha to CSP for vica
VICA uses recaptcha so we need to whitelist google recaptcha urls
in our CSP.

The google developer docs (https://developers.google.com/recaptcha/docs/faq) specifies that the nonce-based approach is recommended,
however since we run static websites we don't have a way of updating
the CSP dynamically. Thus, we are using the alternative specified
way of whitelisting:
- https://www.google.com/recaptcha/
- https://www.gstatic.com/recaptcha/
for script-src and frame-src.